+++ CHANGELOG.md
@@ -6,7 +6,14 @@
 - Added basic driver methods (connection, motor control)
 - Implemented primitive user interfaces
 - Integrated automatic update system
+- Advanced logging system
 
-### Windows (build 2)
-- Added a printer management tab
+### Windows (build 14)
+#### Added
+- A printer management tab
 - Implemented an updater
+
+#### Fixed
+- Installer autorun error after update
+- Error deleting old uninstallers and temporary files
+- Search for Bluetooth Low Energy devices